dronegen fixes from feedback
